Add heuristic to catch libraries that have version in their name.
In some cases a library may contain version in the name.
Unfortunately, CScout does not process such libraries but it may process
the library without the version name.
For example, in libc there are libc.so and libc-2.28.so.
Hence we provide a heuristic that tries to use libc.so if it couldn't
find libc-2.28.so.